* applied patch by kaliber that fixes more than 100 warnings - thanks a lot! Closes #6349
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@37670 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/translators/ico/ICOTranslator.cpp
@@ -159,7 +159,7 @@ ICOTranslator::DerivedTranslate(BPositionIO *source,
 status_t
 ICOTranslator::DerivedCanHandleImageSize(float width, float height) const
 {
-	if (!ICO::is_valid_size(width) || !ICO::is_valid_size(height))
+	if (!ICO::is_valid_size((int)width) || !ICO::is_valid_size((int)height))
 		return B_NO_TRANSLATOR;
 	return B_OK;
 }
